requirements.txt
chromadb
openai
python-dotenv==1.0.1
sentence-transformers
pinecone-client==5.0.1